skip  wasm-gc-breakpoints when stress_incremental_marking

+++ src/baseline/riscv64/baseline-assembler-riscv64-inl.h
@@ -36,7 +36,7 @@ class BaselineAssembler::ScratchRegisterScope {
   UseScratchRegisterScope wrapped_scope_;
 };
 
-enum class Condition : uint8_t {
+enum class Condition : uint32_t {
   kEqual = eq,
   kNotEqual = ne,
